backend/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Migrate==4.0.5
Flask-JWT-Extended==4.5.3
Flask-Cors==3.0.10
python-jose==3.3.0
psycopg2-binary==2.9.7
celery==5.3.4
redis==5.0.1
mediapipe==0.10.3
opencv-python==4.8.0.76
numpy==1.25.2
Werkzeug==2.3.7
python-dotenv==1.0.0
alembic==1.12.0
SQLAlchemy==2.0.23
gunicorn==21.2.0
pytest==7.4.3
pytest-cov==4.1.0
black==23.11.0
flake8==6.1.0
isort==5.12.0
